separate helpers for usecases api
